Restricted Parameters
Using 'lo' and 'hi' to restrict the tuple of parameters and stop the modelfit from predicting negative values for 'β'.

The parameter 'β΄ was minimised by the fit, suggesting that the behaviour of the spring pot can be reduced to a spring. Next, a standard linear solid will be tested (to replace the springpot with a simple spring)